backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn>=0.22.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# HTTP client for Ollama API
httpx>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0

# Message Queue
pika>=1.3.0  # RabbitMQ client
aio-pika>=9.0.0  # Async RabbitMQ client

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0  # Also used for TestClient